Windows does not have "/tmp" path usually

+++ tool/lib/_tmpdir.rb
@@ -2,9 +2,11 @@ require "tmpdir"
 require "fileutils"
 
 template = "rubytest."
-if (tmpdir = Dir.mktmpdir(template)).size > 50
+if (tmpdir = Dir.mktmpdir(template)).size > 50 and File.directory?("/tmp")
   # On macOS, the default TMPDIR is very long, inspite of UNIX socket
   # path length is limited.
+  # On Windows, UNIX socket is not available and no need to shorten
+  # TMPDIR, otherwise assume "/tmp" always exists.
   Dir.rmdir(tmpdir)
   tmpdir = Dir.mktmpdir(template, "/tmp")
 end
